asdad
FIx: add get_time_zone method and enum TimeZoneFormat


Problem:
We're in the GetSystemDateAndTime request according to
http://www.onvif.org/ver10/device/wsdl/devicemgmt.wsdl
should return Timezone information in Posix format.
To be precise according to: ONVIF-Core-Specification-v220.pdf
see paragraph (8.3.6 Get system date and time)
the following is written: "TimeZone": The time zone as it is defined in POSIX 1003.1 section 8.3

> TZ This variable shall represent timezone information. The contents of the
> environment variable named TZ shall be used by the ctime(), ctime_r(),
> localtime(), localtime_r() strftime(), mktime(), functions, and by various
> utilities, to override the default timezone.

see:
 - http://www.open-std.org/jtc1/sc22/open/n4217.pdf
 - https://www.gnu.org/software/libc/manual/html_node/TZ-Variable.html

An Example
   Eastern USA time zone, 1986:

   EST5EDT4,116/02:00:00,298/02:00:00

   "America/Sao_Paulo" rather than specifying the more complex
   TZ=GRNLNDST3GRNLNDDT,M10.3.0/00:00:00,M2.4.0/00:00:00.

We used the tm_zone field from struct tm which was returned by the localtime() function
more information see here:
 - https://www.man7.org/linux/man-pages/man3/localtime.3p.html
 - https://www.man7.org/linux/man-pages/man3/tzset.3p.html

But in most cases it is either just an offset: +0400 or a short abbreviation like std
These values ​​are not CORRECT (Full) values ​​and utilities such as:
ONVIF_device_manager-v2.2.250 or ONVIF_Device_TestTool_v16.01
report that the given value is wrong!

Further we will use abbreviations:
 - DM == ONVIF_device_manager-v2.2.250
 - TT == ONVIF_Device_TestTool_v16.01


Therefore, using tm_zone from struct tm or the global variable tzname is not correct!
Incorrect in terms of full format and ONVIF requirements.

We can use the TZ environment variable. But in most cases it is not installed!
On my Linux Kubuntu 20.04 system the timing is working correctly but TZ is not installed.

We can get TZ by examining the localtime sources and parse the tiz database
This database and sources can be found here: https://www.iana.org/time-zones

But I consider this decision redundant and complicated.

DM or TT can handle a simple offset. in the form: [+|-]hh:mm or UTC[+|-]hh:mm or GMT[+|-]hh:mm

One feature is that DM only handles GMT[+|-]hh:mm correctly
Correct in the sense, it corrects the received time according to the specified offset!
If you specify offset in the form: [+|-]hh:mm or UTC[+|-]hh:mm there is no error!
BUT the time is recalculated in the opposite direction!

I understand that this is due to the sign of the offset!
By definition, GMT and UTC mean the same thing!
But when specifying offsets, the sign of the GMT is inverted!
This feature is described here: https://en.wikipedia.org/wiki/Tz_database#Area

A list of standard offsets according to the TimeZone names is presented here:
https://en.wikipedia.org/wiki/List_of_tz_database_time_zones

As an example: Etc/GMT+9 is UTC+09:00


You also need to consider the various options for the TimeZone database.
For example, Windows uses its own database, while Unix
used by The IANA/Olson Time Zone Database

This issue is discussed here: https://stackoverflow.com/tags/timezone/info


Considering all these factors, it is not correct to use any one option.
Therefore, I added the method `get_time_zone`, which calculates the
TimeZone according to the given format:`TimeZoneFormat`

```
enum TimeZoneForamt: unsigned int
{
    TZ_UTC_OFFSET,  // [+|-]hh:mm
    TZ_GMT_OFFSET,  // [+|-]hh:mm (reverse sign of TZ_UTC_OFFSET)
    TZ_UTC,         // UTC [+|-]hh:mm
    TZ_GMT,         // GMT [+|-]hh:mm (reverse sign of TZ_UTC)
    TZ_ENV,         // value from TZ environment variable
    TZ_ONVIF_ENV,   // value from TZ_ONVIF environment variable (its not std)

    TZ_CNT_FORMATS  // Its not format! Its counter for use in code (max index)
};
```

If none of the formats suits you, and you cannot (or don’t want to)
change the TZ environment variable, then you can use the TZ_ONVIF_ENV format
and set the TZ_ONVIF environment variable

I think this solution is flexible enough to solve the issue with TimeZone and various types of DVRs,
even if they are using a non-standard version of the database.
We can hardcode TZ_ONVIF without harming the Linux system, the daemon is running on.